change tuple to float

+++ src/at/ac/univie/bang/data/DataWorker.java
@@ -8,6 +8,6 @@ public interface DataWorker {
 	
 	public int getCurPosition();
 	
-	public int[] readTuple();
+	public float[] readTuple();
 
 }
